requirements.txt
torch>=1.6.0
adapter-transformers==2.0.1
datasets==1.3.0
numpy==1.18.1
tokenizers==0.10.3
tensorboardX==2.1
tqdm==4.49.0
scikit-learn==0.21.2
